test: split Struct codec input containment coverage
